# api.yaml
openapi: 3.0.3
info:
  title: BOAS 서비스 통합 API
  description: BOAS 공지사항, 채용 및 제품 관리 시스템을 위한 API 명세서입니다.
  version: 1.8.1
servers:
  - url: /api/v1
    description: 기본 API 경로
...
security:
  - {}
paths:
...
  /auth/login:
    post:
      summary: 관리자 로그인
      description: 비밀번호를 검증하여 JWT 액세스 토큰을 발급합니다.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: 로그인 성공 (토큰 발급)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: 인증 실패 (비밀번호 불일치)
  /auth/verify:
    get:
      summary: 토큰 유효성 검증
      description: 헤더에 포함된 JWT 토큰이 유효한지 확인합니다. (페이지 새로고침 시 사용)
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 유효한 토큰
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
        '401':
          description: 유효하지 않은 토큰

  # --- Partner API ---
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      properties:
        password:
          type: string
          description: 관리자 비밀번호
          example: secure_password_123!
      required: [password]
    AuthResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            accessToken: 
              type: string
              description: JWT 액세스 토큰

    # --- Partner Schemas ---
...